refactor(app/displayEntityStore): move project file processing logic to separate file
